python/requirements.txt
requests~=2.31.0
grpcio~=1.54.2
grpcio-tools~=1.54.2